project5_churn/requirements.txt
xgboost==2.0.3
shap==0.45.0
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
streamlit==1.35.0
plotly==5.22.0
openai==1.30.1
python-dotenv==1.0.1
